+++ packages/@orchidui-vue/src/StoreDesign.js
@@ -1,4 +1,4 @@
 import SDMenus from "./StoreDesign/SDMenus/SDMenus.vue";
-import ThemeThumbnail from "./ThemeThumbnail.vue";
+import ThemeThumbnail from "./StoreDesign/ThemeThumbnail.vue";
 
 export { SDMenus, ThemeThumbnail };
