--- packages/@orchidui-vue/src/StoreDesign/ThemeThumbnail/ThemeThumbnail.js
@@ -1,2 +1,0 @@
-import ThemeThumbnail from "./ThemeThumbnail.vue";
-export { ThemeThumbnail };
