+++ packages/@orchidui-vue/src/StoreDesign/ThumbnailSection/ThumbnailSection.vue
